+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - Fix: `PropertiesEndpoint.retrieve` now actually works and either returns a single property item or a list.
 - Chg: Renamed `db.fetch_all()` to `db.get_all_pages()`.
 - New: Query databases with a PySpark/Polars inspired DSL, e.g. `uno.prop('Name') == 'Linus Torvalds'`, implemented.
+- Fix: Database pages created with `db.create_page` are now added to the session cache.
 
 ## Version 0.6, 2024-09-28